LU-6877 osc: cast cl_max_pages_per_rpc to 64bit.
There is a potential overflow as cl_max_pages_per_rpc
is a 32bit type. So cast it to 64bit before shift.

+++ lustre/osc/osc_request.c
@@ -695,11 +695,13 @@ static void osc_shrink_grant_local(struct client_obd *cli, struct obdo *oa)
 static int osc_shrink_grant(struct client_obd *cli)
 {
 	__u64 target_bytes = (cli->cl_max_rpcs_in_flight + 1) *
-			     (cli->cl_max_pages_per_rpc << PAGE_CACHE_SHIFT);
+			     ((__u64)cli->cl_max_pages_per_rpc
+			      << PAGE_CACHE_SHIFT);
 
 	spin_lock(&cli->cl_loi_list_lock);
 	if (cli->cl_avail_grant <= target_bytes)
-		target_bytes = cli->cl_max_pages_per_rpc << PAGE_CACHE_SHIFT;
+		target_bytes = (__u64)cli->cl_max_pages_per_rpc
+			       << PAGE_CACHE_SHIFT;
 	spin_unlock(&cli->cl_loi_list_lock);
 
 	return osc_shrink_grant_to_target(cli, target_bytes);
@@ -715,8 +717,9 @@ int osc_shrink_grant_to_target(struct client_obd *cli, __u64 target_bytes)
 	/* Don't shrink if we are already above or below the desired limit
 	 * We don't want to shrink below a single RPC, as that will negatively
 	 * impact block allocation and long-term performance. */
-	if (target_bytes < cli->cl_max_pages_per_rpc << PAGE_CACHE_SHIFT)
-		target_bytes = cli->cl_max_pages_per_rpc << PAGE_CACHE_SHIFT;
+	if (target_bytes < (__u64)cli->cl_max_pages_per_rpc << PAGE_CACHE_SHIFT)
+		target_bytes = (__u64)cli->cl_max_pages_per_rpc
+			       << PAGE_CACHE_SHIFT;
 
 	if (target_bytes >= cli->cl_avail_grant) {
 		spin_unlock(&cli->cl_loi_list_lock);
@@ -763,7 +766,8 @@ static int osc_should_shrink_grant(struct client_obd *client)
 		/* Get the current RPC size directly, instead of going via:
 		 * cli_brw_size(obd->u.cli.cl_import->imp_obd->obd_self_export)
 		 * Keep comment here so that it can be found by searching. */
-		int brw_size = client->cl_max_pages_per_rpc << PAGE_CACHE_SHIFT;
+		int brw_size = (__u64)client->cl_max_pages_per_rpc
+			       << PAGE_CACHE_SHIFT;
 
 		if (client->cl_import->imp_state == LUSTRE_IMP_FULL &&
 		    client->cl_avail_grant > brw_size)
